USBCAN2MQTT-addon/rootfs/usr/bin/requirements.txt
gs-usb>=0.3.0
paho-mqtt>=2.1.0
python-can>=4.3.0,<5.0.0
pyusb>=1.2.1